requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-filter==2.0.0
django-haystack==2.8.1
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
idna==2.7
Markdown==3.0.1
pysolr==3.8.1
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
